+++ mobile/appcontainer/IdlePanel.js
@@ -5,49 +5,42 @@
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
 
-import ReactDOM from 'react-dom';
 import {XH, hoistCmp} from '@xh/hoist/core';
 import {vframe, div, img, p} from '@xh/hoist/cmp/layout';
 import {panel} from '@xh/hoist/mobile/cmp/panel';
 import {Icon} from '@xh/hoist/icon';
 import {button} from '@xh/hoist/mobile/cmp/button';
 
-import './IdleDialog.scss';
-import idleImage from './IdleDialogImage.png';
+import './IdlePanel.scss';
+import idleImage from './IdlePanelImage.png';
 
 /**
- * Default dialog to display when the app has suspended itself due to inactivity.
- * This display can be overridden by applications - {@see AppSpec.idleDialogClass}. *
+ * Default panel to display when the app has suspended itself due to inactivity.
+ * This display can be overridden by applications - {@see AppSpec.idlePanelClass}. *
  * @private
  */
-export const idleDialog = hoistCmp.factory({
-    displayName: 'IdleDialog',
+export const idlePanel = hoistCmp.factory({
+    displayName: 'IdlePanel',
 
     render({onReactivate}) {
-        return ReactDOM.createPortal(idleDialogPanel({onReactivate}), document.body);
-    }
-});
-
-const idleDialogPanel = hoistCmp.factory(
-    ({onReactivate}) => {
         return panel({
-            className: 'xh-idle-dialog',
+            className: 'xh-idle-panel',
             title: `${XH.clientAppName} is sleeping`,
             icon: Icon.moon(),
             items: [
                 img({src: idleImage}),
                 vframe({
-                    className: 'xh-idle-dialog__content',
+                    className: 'xh-idle-panel__content',
                     items: [
                         div({
-                            className: 'xh-idle-dialog__text-container',
+                            className: 'xh-idle-panel__text-container',
                             items: [
                                 p('This application is sleeping due to inactivity.'),
                                 p('Please click below to reload it.')
                             ]
                         }),
                         div({
-                            className: 'xh-idle-dialog__button-container',
+                            className: 'xh-idle-panel__button-container',
                             item: button({
                                 text: "I'm back!",
                                 flex: 1,
@@ -59,4 +52,4 @@ const idleDialogPanel = hoistCmp.factory(
             ]
         });
     }
-);
+});