+++ svc/IdleService.js
@@ -4,10 +4,10 @@
  *
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
-import {AppState, HoistService, XH} from '@xh/hoist/core';
+import {AppState, HoistService, XH, managed} from '@xh/hoist/core';
 import {Timer} from '@xh/hoist/utils/async';
 import {MINUTES} from '@xh/hoist/utils/datetime';
-import {debounce} from 'lodash';
+import {isNil} from 'lodash';
 
 /**
  * Manage the idling/suspension of this application after a certain period of user inactivity
@@ -16,14 +16,14 @@ import {debounce} from 'lodash';
  * system from unattended clients and/or as a "belt-and-suspenders" defence against memory
  * leaks or other performance issues that can arise with long-running sessions.
  *
- * This service consults the AppSpec `idleDetectionEnabled` property, the `xhIdleTimeoutMins`
- * soft-config, and the `xh.disableIdleDetection` user preference to determine if and when it
- * should suspend the app.
+ * This service consults the `xhIdleTimeouts` soft-config and the `xh.disableIdleDetection`
+ * user preference to determine if and when it should suspend the app.
  */
 @HoistService
 export class IdleService {
 
-    ACTIVITY_EVENTS = ['keydown', 'mousemove', 'mousedown', 'scroll'];
+    @managed timer = null;
+    timeout = null;
 
     constructor() {
         this.addReaction({
@@ -36,36 +36,34 @@ export class IdleService {
     // Implementation
     //------------------------
     startMonitoring() {
-        const timeout = XH.getConf('xhIdleTimeoutMins') * MINUTES,
-            appEnabled = XH.appSpec.idleDetectionEnabled,
+        const timeoutConfigs = XH.getConf('xhIdleTimeouts', {}),
+            timeout = !isNil(timeoutConfigs[XH.clientAppCode]) ? timeoutConfigs[XH.clientAppCode] * MINUTES : -1,
             configEnabled = timeout > 0,
             userEnabled = !XH.getPref('xhIdleDetectionDisabled');
 
-        if (appEnabled && configEnabled && userEnabled) {
-            this.startCountdown = debounce(() => this.suspendApp(), timeout, {trailing: true});
-            this.startCountdown();
-            this.createAppListeners();
+        if (configEnabled && userEnabled) {
+            this.timeout = timeout;
+            this.createTimer();
         }
     }
 
-    createAppListeners() {
-        this.ACTIVITY_EVENTS.forEach(e => {
-            window.addEventListener(e, this.startCountdown, true);
+    createTimer() {
+        this.timer = Timer.create({
+            runFn: () => this.checkInactivityTimeout(),
+            interval: 500
         });
     }
 
-    destroyAppListeners() {
-        this.ACTIVITY_EVENTS.forEach(e => {
-            window.removeEventListener(e, this.startCountdown, true);
-        });
+    checkInactivityTimeout() {
+        if (XH.appState === AppState.SUSPENDED) return;
+        const inactiveTime = Date.now() - XH.lastActivityMs;
+        if (inactiveTime > this.timeout) this.suspendApp();
     }
 
     suspendApp() {
-        if (XH.appState !== AppState.SUSPENDED) {
-            XH.setAppState(AppState.SUSPENDED);
-            this.destroyAppListeners();
-            XH.webSocketService.shutdown();
-            Timer.cancelAll();
-        }
+        if (XH.appState === AppState.SUSPENDED) return;
+        XH.setAppState(AppState.SUSPENDED);
+        XH.webSocketService.shutdown();
+        Timer.cancelAll();
     }
 }